update dao services functions calls
