+++ src/services/networkService.ts
@@ -2188,302 +2188,6 @@ class NetworkService {
   }
 
   /***********************************************/
-  /*****         DAO Functions               *****/
-  /***********************************************/
-
-  // TODO: dao.service.ts
-  // get DAO Balance
-  async getDaoBalance() {
-    if (!this.BobaContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: getDaoBalance() error - called but account === null')
-      return
-    }
-
-    try {
-      const balance = await this.BobaContract!.balanceOf(this.account)
-      return { balance: formatEther(balance) }
-    } catch (error) {
-      console.log('Error: getDaoBalance', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  async getDaoBalanceX() {
-    if (!this.xBobaContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: getDaoBalanceX() error - called but account === null')
-      return
-    }
-
-    try {
-      const balance = await this.xBobaContract.balanceOf(this.account)
-      return { balanceX: formatEther(balance) }
-    } catch (error) {
-      console.log('Error: getDaoBalanceX', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  // get DAO Votes
-  async getDaoVotes() {
-    if (!this.BobaContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: getDaoVotes() error - called but account === null')
-      return
-    }
-
-    try {
-      const votes = await this.BobaContract!.getCurrentVotes(this.account)
-      return { votes: formatEther(votes) }
-    } catch (error) {
-      console.log('NS: getDaoVotes error:', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  // get DAO Votes
-  async getDaoVotesX() {
-    if (!this.xBobaContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: getDaoVotesX() error - called but account === null')
-      return
-    }
-
-    try {
-      const votes = await this.xBobaContract.getCurrentVotes(this.account)
-      return { votesX: formatEther(votes) }
-    } catch (error) {
-      console.log('NS: getDaoVotesX error:', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  //Delegate DAO Authority
-  async delegateVotes({ recipient }) {
-    if (this.L1orL2 !== 'L2') {
-      return
-    }
-    if (!this.BobaContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: delegateVotes() error - called but account === null')
-      return
-    }
-
-    try {
-      const tx = await this.BobaContract!.connect(
-        this.provider!.getSigner()
-      ).delegate(recipient)
-      await tx.wait()
-      return tx
-    } catch (error) {
-      console.log('NS: delegateVotes error:', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  //Delegate DAO Authority
-  async delegateVotesX({ recipient }) {
-    if (this.L1orL2 !== 'L2') {
-      return
-    }
-    if (!this.xBobaContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: delegateVotesX() error - called but account === null')
-      return
-    }
-
-    try {
-      const tx = await this.xBobaContract
-        .connect(this.provider!.getSigner())
-        .delegate(recipient)
-      await tx.wait()
-      return tx
-    } catch (error) {
-      console.log('NS: delegateVotesX error:', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  // Proposal Create Threshold
-  async getProposalThreshold() {
-    if (!this.delegateContract) {
-      return
-    }
-
-    try {
-      const delegateCheck = await this.delegateContract.attach(
-        this.addresses.GovernorBravoDelegator
-      )
-      const rawThreshold = await delegateCheck.proposalThreshold()
-      return { proposalThreshold: formatEther(rawThreshold) }
-    } catch (error) {
-      console.log('NS: getProposalThreshold error:', error)
-      return error
-    }
-  }
-
-  // Create Proposal
-  /************************/
-  /*****Old Dao Fix Me.****/
-  /************************/
-  // TODO: dao.service.ts
-  // FIXME:
-  async createProposal(payload) {
-    if (this.L1orL2 !== 'L2') {
-      return
-    }
-    if (!this.delegateContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: delegateVotesX() error - called but account === null')
-      return
-    }
-
-    let signatures = ['']
-    let value1 = 0
-    let value2 = 0
-    let value3 = 0
-    let description = ''
-    let address = ['']
-    let callData = ['']
-    // FIXME: Ve DAO From here
-    /*
-      let tokenIds = payload.tokenIds
-      // create proposal only on latest contracts.
-      const delegateCheck = await this.delegateContract.attach(this.addresses.GovernorBravoDelegatorV2)
-
-    */
-    // FIXME: Ve DAO Till here
-
-    const delegateCheck = await this.delegateContract.attach(
-      this.addresses.GovernorBravoDelegator
-    )
-
-    if (payload.action === 'text-proposal') {
-      address = ['0x000000000000000000000000000000000000dEaD']
-      description = payload.text.slice(0, 252) //100+150+2
-      callData = [
-        '0x0000000000000000000000000000000000000000000000000000000000000000',
-      ]
-    } else if (payload.action === 'change-lp1-fee') {
-      signatures = ['configureFeeExits(uint256,uint256,uint256)']
-      value1 = Number(payload.value[0])
-      value2 = Number(payload.value[1])
-      value3 = Number(payload.value[2])
-      description = `Change L1 LP Bridge fee to ${value1}, ${value2}, and ${value3} integer percent`
-      address = [this.addresses.L2LPAddress]
-      callData = [
-        ethers.utils.defaultAbiCoder.encode(
-          ['uint256', 'uint256', 'uint256'],
-          [value1, value2, value3]
-        ),
-      ]
-    } else if (payload.action === 'change-lp2-fee') {
-      address = [delegateCheck.address]
-      signatures = ['configureFee(uint256,uint256,uint256)']
-      value1 = Number(payload.value[0])
-      value2 = Number(payload.value[1])
-      value3 = Number(payload.value[2])
-      description = `Change L2 LP Bridge fee to ${value1}, ${value2}, and ${value3} integer percent`
-      address = [this.addresses.L2LPAddress]
-      callData = [
-        ethers.utils.defaultAbiCoder.encode(
-          ['uint256', 'uint256', 'uint256'],
-          [value1, value2, value3]
-        ),
-      ]
-    } else if (payload.action === 'change-threshold') {
-      address = [delegateCheck.address]
-      signatures = ['_setProposalThreshold(uint256)']
-      value1 = Number(payload.value[0])
-      description = `Change Proposal Threshold to ${value1} BOBA`
-      callData = [ethers.utils.defaultAbiCoder.encode(['uint256'], [value1])]
-    }
-
-    try {
-      const values = [0] //amount of ETH to send, generally, zero
-
-      return await delegateCheck
-        .connect(this.provider!.getSigner())
-        .propose(address, values, signatures, callData, description)
-    } catch (error) {
-      console.log('NS: createProposal error:', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  //Fetch DAO Proposals
-  async queueProposal(proposalID) {
-    if (!this.delegateContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: queueProposal() error - called but account === null')
-      return
-    }
-
-    try {
-      const delegateCheck = this.delegateContract
-        .connect(this.provider!.getSigner())
-        .attach(this.addresses.GovernorBravoDelegator)
-      return delegateCheck.queue(Number(proposalID))
-    } catch (error) {
-      console.log('NS: queueProposal error:', error)
-      return error
-    }
-  }
-  // TODO: dao.service.ts
-  async executeProposal(proposalID) {
-    if (!this.delegateContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: executeProposal() error - called but account === null')
-      return
-    }
-
-    try {
-      const delegateCheck = this.delegateContract
-        .connect(this.provider!.getSigner())
-        .attach(this.addresses.GovernorBravoDelegator)
-      return delegateCheck.execute(Number(proposalID))
-    } catch (error) {
-      console.log('NS: executeProposal error:', error)
-      return error
-    }
-  }
-
-  /***********************************************/
   /*****            L1 Security Fee          *****/
   /***********************************************/
 
@@ -2532,123 +2236,6 @@ class NetworkService {
     )
     return ethers.utils.formatEther(await L2BillingContract.exitFee())
   }
-
-  /****************************************
-   ************* DAO ***********
-   *****************************************/
-  // TODO: dao.service.ts
-  async fetchProposals() {
-    if (!this.delegateContract) {
-      return
-    }
-
-    const delegateCheck = this.delegateContract.attach(
-      this.addresses.GovernorBravoDelegator
-    )
-
-    try {
-      const proposalList: any = []
-
-      const proposalCounts = await delegateCheck.proposalCount()
-      const totalProposals = await proposalCounts.toNumber()
-
-      /// @notice An event emitted when a new proposal is created
-      // event ProposalCreated(uint id, address proposer, address[] targets, uint[] values, string[] signatures, bytes[] calldatas, uint startTimestamp, uint endTimestamp, string description);
-      const L2ChainId = this.networkConfig!.L2.chainId
-      const descriptionList = await graphQLService.queryBridgeProposalCreated({
-        sourceChainId: L2ChainId,
-      })
-
-      for (let i = 0; i < totalProposals; i++) {
-        const proposalRaw = descriptionList!.data.governorProposalCreateds[i]
-
-        if (typeof proposalRaw === 'undefined') {
-          continue
-        }
-
-        const proposalID = proposalRaw.proposalId
-
-        //this is a number such as 2
-        const proposalData = await delegateCheck.proposals(proposalID)
-
-        const proposalStates = [
-          'Pending',
-          'Active',
-          'Canceled',
-          'Defeated',
-          'Succeeded',
-          'Queued',
-          'Expired',
-          'Executed',
-        ]
-
-        const state = await delegateCheck.state(proposalID)
-
-        const againstVotes = parseInt(
-          formatEther(proposalData.againstVotes),
-          10
-        )
-        const forVotes = parseInt(formatEther(proposalData.forVotes), 10)
-        const abstainVotes = parseInt(
-          formatEther(proposalData.abstainVotes),
-          10
-        )
-
-        const startTimestamp = proposalData.startTimestamp.toString()
-        const endTimestamp = proposalData.endTimestamp.toString()
-
-        const proposal = await delegateCheck.getActions(i + 2)
-
-        const hasVoted = null
-
-        const description = proposalRaw.description.toString()
-
-        proposalList.push({
-          id: proposalID?.toString(),
-          proposal,
-          description,
-          totalVotes: forVotes + againstVotes,
-          forVotes,
-          againstVotes,
-          abstainVotes,
-          state: proposalStates[state],
-          startTimestamp,
-          endTimestamp,
-          hasVoted,
-        })
-      }
-      return { proposalList }
-    } catch (error) {
-      console.log('NS: fetchProposals error:', error)
-      return error
-    }
-  }
-
-  // TODO: dao.service.ts
-  async castProposalVote({ id, userVote }) {
-    if (!this.delegateContract) {
-      return
-    }
-
-    if (!this.account) {
-      console.log('NS: castProposalVote() error - called but account === null')
-      return
-    }
-    try {
-      const delegateCheck = this.delegateContract
-        .connect(this.provider!.getSigner())
-        .attach(this.addresses.GovernorBravoDelegator)
-      return delegateCheck.castVote(id, userVote)
-    } catch (error) {
-      console.log('NS: castProposalVote error:', error)
-      return error
-    }
-  }
-
-  /****************************************
-   ************* END HERE *****************
-   ***********OLD DAO REMOVE ME TILL HERE *
-   *****************************************/
 
   async submitTxBuilder(
     contract: Contract,
